+++ src/main/java/com/ldb/truck/Entity/OrderItem/OrderReqItem.java
@@ -5,19 +5,21 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderReqItem {
     private int item;
-    private int qty;
+    private BigDecimal qty;
     private int realQty;
 
     private String currency;
     private int exchangeRate;
     private Float realPrice;
-    private Float price;
+    private BigDecimal price;
     private int shopeId;
     private String orderType;
     private String placeBuy;
