+++ src/main/java/com/ldb/truck/Repository/ItemEntityRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.security.Timestamp;
 import java.util.Date;
 import java.util.List;
@@ -158,5 +159,5 @@ public interface ItemEntityRepository extends CrudRepository<ItemEntity,Long> {
     @Query(value = "UPDATE item_inventory i SET i.qty = i.qty - :qty  " +
             "WHERE i.item_id = :itemId  ",
             nativeQuery = true)
-    int updateStockInItemOut(Integer qty,Integer itemId);
+    int updateStockInItemOut(BigDecimal qty, Integer itemId);
 }
